requirements.txt
Django==1.8.3
django-filter==0.10.0
django-suit==0.2.14
djangorestframework==3.2.1
Markdown==2.6.2
mock==1.3.0
pbr==1.5.0
Pillow==2.9.0
requests==2.7.0
six==1.9.0
testfixtures==4.2.0
